# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/mcps": {
      "get": {
        "operationId": "listMcps",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomMcpDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the current user's configured MCPs",
        "tags": [
          "mcps"
        ]
      },
      "post": {
        "operationId": "addMcp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCustomMcpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomMcpDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Register a new local MCP for the current user",
        "tags": [
          "mcps"
        ]
      }
    },
    "/mcps/{id}": {
      "patch": {
        "operationId": "updateMcp",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomMcpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomMcpDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update an existing MCP (port, active, headers)",
        "tags": [
          "mcps"
        ]
      },
      "delete": {
        "operationId": "removeMcp",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove an MCP from the current user account",
        "tags": [
          "mcps"
        ]
      }
    },
    "/mcp/logs": {
      "get": {
        "operationId": "listAllMcpLogs",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "1-based page number",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 25,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpLogPageDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List (paginated) proxied request/response logs for every MCP the current user owns",
        "tags": [
          "mcp-logs"
        ]
      },
      "delete": {
        "operationId": "deleteMcpLogs",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteMcpLogsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Number of deleted entries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "number"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Bulk-delete log entries: by id, by mcpId, or every log the caller owns if neither is given",
        "tags": [
          "mcp-logs"
        ]
      }
    },
    "/mcp/{mcpId}/logs": {
      "get": {
        "operationId": "listMcpLogs",
        "parameters": [
          {
            "name": "mcpId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "1-based page number",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 25,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpLogPageDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List (paginated) proxied request/response logs for a single owned MCP",
        "tags": [
          "mcp-logs"
        ]
      }
    },
    "/mcp/logs/{id}": {
      "delete": {
        "operationId": "deleteMcpLog",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a single log entry",
        "tags": [
          "mcp-logs"
        ]
      }
    },
...
  },
  "info": {
    "title": "MCP Bridge API",
    "description": "Relay backend exposing desktop-agent MCP servers over HTTP",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CustomMcpDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "example": "notes",
            "description": "Unique (per user) name — the public id becomes \"<username>-<name>\""
          },
          "port": {
            "type": "number",
            "example": 3333,
            "description": "Port the local MCP server listens on (on the user's machine)"
          },
          "subPath": {
            "type": "string",
            "example": "/api/mcp",
            "description": "Optional sub-path on the local server, e.g. \"/api/mcp\" reaches http://localhost:<port>/api/mcp"
          },
          "active": {
            "type": "boolean"
          },
          "headers": {
            "type": "object",
            "description": "Custom HTTP headers the agent should attach when calling the local MCP server"
          }
        },
        "required": [
          "id",
          "name",
          "port",
          "active"
        ]
      },
      "AddCustomMcpDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "notes"
          },
          "port": {
            "type": "number",
            "example": 3333
          },
          "subPath": {
            "type": "string",
            "example": "/api/mcp",
            "description": "Optional sub-path on the local server"
          },
          "headers": {
            "type": "object"
          }
        },
        "required": [
          "name",
          "port"
        ]
      },
      "UpdateCustomMcpDto": {
        "type": "object",
        "properties": {
          "port": {
            "type": "number",
            "example": 3333
          },
          "subPath": {
            "type": "string",
            "example": "/api/mcp",
            "description": "Optional sub-path on the local server — send \"\" to clear it"
          },
          "active": {
            "type": "boolean"
          },
          "headers": {
            "type": "object"
          }
        }
      },
      "McpLogWithContextDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "requestId": {
            "type": "string",
            "description": "Correlation id for this proxied request, distinct per attempt"
          },
          "method": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "status": {
            "type": "number",
            "description": "HTTP status returned by the agent — absent if the request failed before a response arrived"
          },
          "ok": {
            "type": "boolean",
            "description": "Whether the request completed with a response, regardless of status code"
          },
          "errorMessage": {
            "type": "string",
            "description": "Set when the request failed (agent offline, timeout, ...)"
          },
          "durationMs": {
            "type": "number"
          },
          "requestBody": {
            "type": "object",
            "description": "Body sent to the local MCP server, if any"
          },
          "responseBody": {
            "type": "object",
            "description": "Body returned by the local MCP server — absent if the request failed before a response arrived"
          },
          "timestamp": {
            "type": "string",
            "description": "ISO 8601 timestamp"
          },
          "mcpId": {
            "type": "string"
          },
          "mcpName": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "requestId",
          "method",
          "path",
          "ok",
          "durationMs",
          "timestamp",
          "mcpId",
          "mcpName"
        ]
      },
      "McpLogPageDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/McpLogWithContextDto"
            }
          },
          "total": {
            "type": "number",
            "description": "Total number of matching log entries across all pages"
          },
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "pageSize"
        ]
      },
      "DeleteMcpLogsDto": {
        "type": "object",
        "properties": {
          "ids": {
            "description": "Delete exactly these log entries (bulk-select in the UI)",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "mcpId": {
            "type": "string",
            "description": "Delete every log entry for this one MCP"
          }
        }
      }
    }
  }
}